build: disable minification and externalize all packages
Minification has been disabled to make it easier for others to modify the package code, such as making patches or fixing other issues.

The packages were externalized in order not to load more data again, because the dependencies will be loaded anyway, so there is no point in additionally bundling everything into one file.