feat: ✨ add the ability to enable a buttons to instantly copy or download a page in Markdown format on the human documentation side
I thought it wouldn't be possible to do this (without working on the VitePress code itself), but it turned out that it is possible, although it looks like a crutch, but it works!

Such buttons were added a long time ago in https://bun.sh/docs and on other sites along with `llms.txt`, but it was so lacking in VitePress

Now we have them too! 🔥