refactor: ♻️ refactor the "Copy as Markdown" button components and convert them to a Vue component
The previous solution presented in 4bb73e9268b40fcf494ab0cce1825e9cecec92b9 is bad because all the HTML / CSS / logic was in one `.ts` file and this makes maintenance difficult
More precisely, the formatting itself was not

With Vue, everything is much simpler and clearer.

The logic for configuring these buttons for VitePress has also been changed, now a special `markdown-it` plugin inserts the component into Markdown (a bit of hacking), and the user must also register the global component `<CopyOrDownloadAsMarkdownButtons />` in order for all this to be displayed.

The method using a special `markdown-it` plugin for inserting a component with buttons was written here - https://discord.com/channels/325477692906536972/790509637598314527/1387290658531708929

Co-authored-by: Divyansh Singh <40380293+brc-dd@users.noreply.github.com>