tests: correctly isolate mocks that caused tests to fail in CI
I couldn't figure out what the problem was for a few days, at first I thought it was a problem with Bun (`bun test`), then I thought it was most likely a problem with the `gray-matter` cache (and there were such problems before), I tried to test it by removing caching from the package (via patch), but it didn't help at all, the reason was something else

Then after some time of debugging I thought that there was a problem between the tests, like when I tested only `tests/plugin/index.test.ts` with `bun test` there were no errors at all, it means that the problem was caused by some other tests, and the mocks were poorly isolated and they were also passed to `tests/plugin/index.test.ts`, which is not desirable

Now I fixed the mocks in that file and isolated it with `afterAll(() => readFile.mockReset())` so that the mock is not passed to other tests, and now it works!

By the way, this bug was discovered thanks to a bug (or maybe not) in `bun test`, in which for some unknown reason the order of tests in GitHub Actions on Windows was different.

[test windows]