fix: fix the problem with the VitePress warning `"node:path" has been externalized for browser compatibility`
The `cleanUrl` function has been moved to the `src/utils/shared/index.ts` file where there is no longer a `node:path` import (which was causing the error) and it is now completely runtime agnostic.

Fixes the problem described in #65

Commit is not signed beacause GPG keys is not synced with my Chromebook.

Reported-by: bqx <boqingxin14@gmail.com>