refactor: use local `cspell` config schema from `node_modules`
Because why download the schema from the internet again if it is available locally?

This will improve the hints, for example if we use `cspell` 1.2.3, and the schema is loaded for the latest version (1.3.9), then it can hint parameters that do not exist for us but are in the latest version, which we do not need.